New Output: adding MethodOutput and BlockTrace
Mesh

  - Added BlockTrace class to aid traversing distributed
    array-of-octrees partition (range of elements in root-block array)
  - Added [pr]_method_output_<foo>() entry methods for MethodOutput

Parameters

  - Added initial_hdf5_* parameters for InitialHdf5
  - Renamed method_refresh_<foo> to method_<foo>
  - Added method_output_blocking parameter to define scope of
    domain for each output file

Testing

  - Added test_BlockTrace unit tests
  - Removed unused code from test/index.php

Method / Output

  - Added MethodOutput "method" for writing HDF5 output

Cello

  - fixed LOAD_ARRAY_TYPE()

Charm

  - Added MsgOutput Charm++ message for use by MethodOutput
  - cleaned MsgRefresh: removed debug code

Initial

  - Added InitialHdf5 stub

Cleaning

  - addressed some -Wall compiler warnings

+++ src/Cello/control_output.cpp
@@ -87,8 +87,8 @@ void Problem::new_output_next(Simulation * simulation) throw()
   // print blocks on this process to check block_vec_ array
 
   Hierarchy * hierarchy = simulation->hierarchy();
-  for (int i=0; i<hierarchy->num_blocks(); i++) {
-    CkPrintf ("%d Block %d = %s\n",CkMyPe(),i,
+  for (size_t i=0; i<hierarchy->num_blocks(); i++) {
+    CkPrintf ("%d Block %u = %s\n",CkMyPe(),i,
 	      hierarchy->block(i)->name().c_str());
   }
   if (output != NULL) {
